test: simple arithmetic output
In both python and assembly
