+++ tests/execution.rs
@@ -3,79 +3,27 @@
 //! equivalent to a reference file, these actually run the compiled output and assert on what it
 //! prints -- for both backends.
 
+use std::error::Error;
 use std::path::Path;
-use std::process::Command;
 
-use assert_cmd::prelude::*;
-use mamba::backend::Backend;
-use mamba::{transpile_dir, Arguments};
-use tests_util::{resource_path, run_python};
+use test_case::test_case;
+use tests_util::{resource_path, run_cli, run_via_bin, run_via_python, Runner};
 
-#[test]
-fn python_backend_prints_expected_output() -> Result<(), Box<dyn std::error::Error>> {
-    let src_dir = resource_path(true, &["function"], "");
-    let out_dir = tempfile::tempdir()?;
-
-    let arguments = Arguments::default(); // backend defaults to `Backend::Python`
-    let output_dir = transpile_dir(
-        Path::new(&src_dir),
-        Some("hello_world.mamba"),
-        Some(out_dir.path().join("out").to_str().unwrap()),
-        &arguments,
-    )
-    .map_err(|errs| format!("{errs:?}"))?;
-
-    let stdout = run_python(&output_dir.join("hello_world.py"))?;
-    assert_eq!(stdout, "hello world\n");
-    Ok(())
+#[test_case(run_via_python, &["function"], "hello_world.mamba" => "hello world\n")]
+#[test_case(run_via_bin, &["function"], "hello_world.mamba" => "hello world\n")]
+#[test_case(run_via_python, &["function"], "arithmetic_sum.mamba" => "14\n")]
+#[test_case(run_via_bin, &["function"], "arithmetic_sum.mamba" => "14\n")]
+#[test_case(run_via_python, &["function"], "for_loop_sum.mamba" => "10\n")]
+#[test_case(run_via_bin, &["function"], "for_loop_sum.mamba" => "10\n")]
+fn execution(run: Runner, dirs: &[&str], file: &str) -> String {
+    run(dirs, file).unwrap()
 }
 
 #[test]
-fn bin_backend_prints_expected_output() -> Result<(), Box<dyn std::error::Error>> {
-    let src_dir = resource_path(true, &["function"], "");
-    let out_dir = tempfile::tempdir()?;
-    let bin_path = out_dir.path().join("hello_world_bin");
-
-    let arguments = Arguments {
-        annotate: false,
-        backend: Backend::Bin { target: None },
-    };
-    let produced = transpile_dir(
-        Path::new(&src_dir),
-        Some("hello_world.mamba"),
-        Some(bin_path.to_str().unwrap()),
-        &arguments,
-    )
-    .map_err(|errs| format!("{errs:?}"))?;
-
-    let output = Command::new(&produced).output()?;
-    assert!(
-        output.status.success(),
-        "executable exited with an error:\n{}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert_eq!(String::from_utf8(output.stdout)?, "hello world\n");
-    Ok(())
-}
-
-#[test]
-fn asm_backend_prints_disassembly_to_stdout() -> Result<(), Box<dyn std::error::Error>> {
+fn asm_prints_disassembly_to_stdout() -> Result<(), Box<dyn Error>> {
     let src_dir = resource_path(true, &["function"], "");
 
-    let mut cmd = Command::main_binary()?;
-    let output = cmd
-        .current_dir(&src_dir)
-        .arg("--asm")
-        .arg("-i")
-        .arg("hello_world.mamba")
-        .output()?;
-
-    assert!(
-        output.status.success(),
-        "mamba --asm exited with an error:\n{}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-    let stdout = String::from_utf8(output.stdout)?;
+    let stdout = run_cli(Path::new(&src_dir), &["--asm", "-i", "hello_world.mamba"])?;
     assert!(
         stdout.contains("; -- main --") && stdout.contains("ret"),
         "expected disassembly on stdout, got:\n{stdout}"
