+++ tests/resource/valid/function/arithmetic_sum.mamba
@@ -1,0 +1,1 @@
+print(10 + 4)
