+++ tests/resource/valid/function/for_loop_sum.mamba
@@ -1,0 +1,3 @@
+def total := 0
+for i in 0 ..= 4 do total := total + i end
+print(total)
