+++ tests_util/Cargo.toml
@@ -10,3 +10,4 @@ mamba = { path = ".." }
 python-parser = "0.1.0"
 itertools = "0.8.2"
 tempfile = "3.1.0"
+assert_cmd = "0.10"
